tabpro-common: Add our own power HAL
The CM PowerHAL makes decisions that override what we configure
ourselves for the governor settings.  That's not nice.

There is too much CPU boosting going on with our Android stack.
Userspace boosting is a little bit all over the place (some
boosts are >=500ms which makes no sense) and overly aggressive.

Let's trust the governor to scale properly and the cpu-boost
driver to handle instantaneous boosting.

Change-Id: I0e07d68a0acf3d78c09b7a5e622d45236fe91b13