tabpro-common: overlay: Allow physical keyboards
This overlay forces the system to believe that there isn't a
physical keyboard.  Bluetooth keyboards are considered
physical keyboards and therefore, prior to this change,
are only partially supported.

In particular, this causes Settings to allow you to configure
the physical keyboard and allows you to use the keyboard without
also having the soft keyboard active on the device at the same
time.

Thanks to xda user Barracuda77777 for showing me this.